Moved openForm(),openSavedForm and prefillAndOpenForm() to FormsInteractor
