+++ odk/extension/src/main/java/io/samagra/odk/collect/extension/handlers/ODKFormsHandler.kt
@@ -3,8 +3,12 @@ package io.samagra.odk.collect.extension.handlers
 import android.content.Context
 import android.content.Intent
 import android.util.Log
+import io.reactivex.rxjava3.disposables.CompositeDisposable
+import io.samagra.odk.collect.extension.interactors.FormInstanceInteractor
 import io.samagra.odk.collect.extension.interactors.FormsDatabaseInteractor
 import io.samagra.odk.collect.extension.interactors.FormsInteractor
+import io.samagra.odk.collect.extension.interactors.FormsNetworkInteractor
+import io.samagra.odk.collect.extension.listeners.FileDownloadListener
 import io.samagra.odk.collect.extension.listeners.FormsProcessListener
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
@@ -12,6 +16,7 @@ import kotlinx.coroutines.launch
 import org.javarosa.core.model.FormDef
 import org.odk.collect.android.activities.FormEntryActivity
 import org.odk.collect.android.events.FormEventBus
+import org.odk.collect.android.events.FormStateEvent
 import org.odk.collect.android.external.FormsContract
 import org.odk.collect.android.formentry.loading.FormInstanceFileCreator
 import org.odk.collect.android.formentry.saving.DiskFormSaver
@@ -24,6 +29,8 @@ import org.odk.collect.android.utilities.ApplicationConstants
 import org.odk.collect.android.utilities.MediaUtils
 import org.odk.collect.entities.EntitiesRepository
 import org.odk.collect.forms.Form
+import org.odk.collect.forms.instances.Instance
+import org.odk.collect.forms.instances.InstancesRepository
 import org.w3c.dom.Document
 import java.io.File
 import java.io.FileOutputStream
@@ -39,8 +46,14 @@ class ODKFormsHandler @Inject constructor(
     private val formsDatabaseInteractor: FormsDatabaseInteractor,
     private val storagePathProvider: StoragePathProvider,
     private val mediaUtils: MediaUtils,
-    private val entitiesRepository: EntitiesRepository
+    private val entitiesRepository: EntitiesRepository,
+    private val formsNetworkInteractor: FormsNetworkInteractor,
+    private val formsInteractor: FormsInteractor,
+    private val instancesRepository: InstancesRepository,
+    private val formInstanceInteractor: FormInstanceInteractor
 ): FormsInteractor {
+
+
 
     override fun openFormWithFormId(formId: String, context: Context) {
         val form = formsDatabaseInteractor.getLatestFormById(formId)
@@ -187,5 +200,117 @@ class ODKFormsHandler @Inject constructor(
             return document
         }
         return document
+    }
+
+    override fun openForm(formId: String, context: Context) {
+        CoroutineScope(Job()).launch {
+            // Delete any saved instances of this form
+            val savedInstances = instancesRepository.getAllByFormId(formId)
+            for (instance in savedInstances) {
+                if (instance.status == Instance.STATUS_INCOMPLETE) {
+                    instancesRepository.delete(instance.dbId)
+                }
+            }
+            val requiredForm = formsDatabaseInteractor.getLatestFormById(formId)
+            if (requiredForm == null) {
+                downloadAndOpenForm(formId, context)
+            }
+            else {
+                val xmlFile = File(requiredForm.formFilePath)
+                if (xmlFile.exists() && (requiredForm.formMediaPath == null || mediaExists(requiredForm))) {
+                    formsInteractor.openFormWithFormId(formId, context)
+                }
+                else {
+                    requiredForm.formMediaPath?.let { File(it).deleteRecursively() }
+                    xmlFile.delete()
+                    formsDatabaseInteractor.deleteByFormId(formId)
+                    downloadAndOpenForm(formId, context)
+                }
+            }
+        }
+    }
+
+    override fun openSavedForm(formId: String, context: Context) {
+        CoroutineScope(Job()).launch {
+            val compositeDisposable = CompositeDisposable()
+            compositeDisposable.add(
+                FormEventBus.getState()
+                    .subscribe { event ->
+                        when (event) {
+                            is FormStateEvent.OnFormOpenFailed -> {
+                                if (event.formId == formId) {
+                                    compositeDisposable.clear()
+                                    openForm(formId, context)
+                                }
+                            }
+                            is FormStateEvent.OnFormOpened -> {
+                                if (event.formId == formId) {
+                                    compositeDisposable.clear()
+                                }
+                            }
+                            else -> {}
+                        }
+                    }
+            )
+
+            formInstanceInteractor.openLatestSavedInstanceWithFormId(formId, context)
+        }
+    }
+
+    override fun prefillAndOpenForm(formId: String, tagValueMap: HashMap<String, String>, context: Context) {
+        CoroutineScope(Job()).launch {
+            val compositeDisposable = CompositeDisposable()
+            compositeDisposable.add(FormEventBus.getState().subscribe { event ->
+                when (event) {
+                    is FormStateEvent.OnFormSaved -> {
+                        if (event.formId == formId) {
+                            val prefilledInstance = formInstanceInteractor.getInstanceByPath(event.instancePath)
+                            if (prefilledInstance != null) {
+                                formInstanceInteractor.openInstance(prefilledInstance, context)
+                            }
+                            else {
+                                FormEventBus.formOpenFailed(formId, "Form instance cannot be found!")
+                            }
+                            compositeDisposable.clear()
+                        }
+                    }
+                    is FormStateEvent.OnFormOpenFailed -> if (event.formId == formId) compositeDisposable.clear()
+                    is FormStateEvent.OnFormSaveError -> if (event.formId == formId) compositeDisposable.clear()
+                    else -> {}
+                }
+            })
+            formsInteractor.prefillForm(formId, tagValueMap)
+        }
+    }
+
+
+
+    private fun downloadAndOpenForm(formId: String, context: Context) {
+        formsNetworkInteractor.downloadFormById(formId, object : FileDownloadListener {
+            override fun onComplete(downloadedFile: File) {
+                formsInteractor.openFormWithFormId(formId, context)
+            }
+        })
+    }
+
+    private fun mediaExists(form: Form): Boolean {
+        val document = DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(File(form.formFilePath))
+        val values = document.getElementsByTagName("value")
+        for (index in 0 until values.length) {
+            val attributes = values.item(index).attributes
+            if (attributes.length > 0) {
+                val nodeValue = attributes.item(0).nodeValue
+                if (nodeValue == "image" || nodeValue == "audio" || nodeValue == "video") {
+                    var mediaFileName = values.item(index).firstChild.nodeValue
+                    if (mediaFileName.isNotBlank()) {
+                        mediaFileName = mediaFileName.substring(mediaFileName.lastIndexOf("/") + 1)
+                        val mediaFile = File(form.formMediaPath + "/" + mediaFileName)
+                        if (!mediaFile.exists())
+                            return false
+                    }
+                }
+            }
+        }
+        return true
     }
 }
