+++ odk/extension/src/main/java/io/samagra/odk/collect/extension/modules/FormsInteractorModule.kt
@@ -4,7 +4,10 @@ import android.app.Application
 import dagger.Module
 import dagger.Provides
 import io.samagra.odk.collect.extension.annotations.ODKFormsInteractor
+import io.samagra.odk.collect.extension.components.DaggerFormInstanceInteractorComponent
 import io.samagra.odk.collect.extension.components.DaggerFormsDatabaseInteractorComponent
+import io.samagra.odk.collect.extension.components.DaggerFormsInteractorComponent
+import io.samagra.odk.collect.extension.components.DaggerFormsNetworkInteractorComponent
 import io.samagra.odk.collect.extension.handlers.ODKFormsHandler
 import io.samagra.odk.collect.extension.interactors.FormsInteractor
 import org.odk.collect.android.injection.config.DaggerAppDependencyComponent
@@ -23,6 +26,11 @@ class FormsInteractorModule {
         val storagePathProvider = appDependencyComponent.storagePathProvider()
         val entitiesRepository = appDependencyComponent.entitiesRepositoryProvider().get(currentProjectProvider.getCurrentProject().uuid)
         val formsDatabaseInteractor = DaggerFormsDatabaseInteractorComponent.factory().create(application).getFormsDatabaseInteractor()
-        return ODKFormsHandler(currentProjectProvider, formsDatabaseInteractor, storagePathProvider, mediaUtils, entitiesRepository)
+        val formsNetworkInteractor = DaggerFormsNetworkInteractorComponent.factory().create(application).getFormsNetworkInteractor()
+        val formsInteractor = DaggerFormsInteractorComponent.factory().create(application).getFormsInteractor()
+        val instancesRepository = DaggerAppDependencyComponent.builder().application(application).build().instancesRepositoryProvider().get()
+        val formInstanceInteractor = DaggerFormInstanceInteractorComponent.factory().create(application).getFormInstanceInteractor()
+
+            return ODKFormsHandler(currentProjectProvider, formsDatabaseInteractor, storagePathProvider, mediaUtils, entitiesRepository,formsNetworkInteractor,formsInteractor,instancesRepository,formInstanceInteractor)
     }
 }
