+++ sample/src/main/java/io/samagra/oce_sample/ODKFeatureTesterActivity.kt
@@ -10,8 +10,10 @@ import android.widget.EditText
 import android.widget.ProgressBar
 import android.widget.Toast
 import io.reactivex.rxjava3.disposables.CompositeDisposable
+import io.samagra.odk.collect.extension.annotations.ODKFormsInteractor
 
 import io.samagra.odk.collect.extension.interactors.FormsDatabaseInteractor
+import io.samagra.odk.collect.extension.interactors.FormsInteractor
 import io.samagra.odk.collect.extension.interactors.FormsNetworkInteractor
 import io.samagra.odk.collect.extension.interactors.ODKInteractor
 import io.samagra.odk.collect.extension.listeners.FileDownloadListener
@@ -40,6 +42,7 @@ class ODKFeatureTesterActivity : AppCompatActivity(), View.OnClickListener {
     private lateinit var progressBar: ProgressBar
 
     private lateinit var odkInteractor: ODKInteractor
+    private lateinit var formsInteractor: FormsInteractor
     private lateinit var networkInteractor: FormsNetworkInteractor
     private lateinit var formsDatabaseInteractor: FormsDatabaseInteractor
 
@@ -67,6 +70,7 @@ class ODKFeatureTesterActivity : AppCompatActivity(), View.OnClickListener {
 
         ODKProvider.init(application)
         odkInteractor = ODKProvider.getOdkInteractor()
+        formsInteractor = ODKProvider.getFormsInteractor()
         progressBar.visibility = View.VISIBLE
         odkInteractor.setupODK(IOUtils.toString(resources.openRawResource(R.raw.settings)), false, object :
             ODKProcessListener {
@@ -118,7 +122,7 @@ class ODKFeatureTesterActivity : AppCompatActivity(), View.OnClickListener {
                 val formId: String = openFormsInput.text.toString().trim()
                 if (formId.isNotBlank()) {
                     progressBar.visibility = View.VISIBLE
-                    odkInteractor.openForm(formId, context)
+                    formsInteractor.openForm(formId, context)
                 }
             }
             R.id.download_form_button -> {
@@ -184,7 +188,7 @@ class ODKFeatureTesterActivity : AppCompatActivity(), View.OnClickListener {
                 val formId: String = openSavedInput.text.toString().trim()
                 if (formId.isNotBlank()) {
                     progressBar.visibility = View.VISIBLE
-                    odkInteractor.openSavedForm(formId, context)
+                    formsInteractor.openSavedForm(formId, context)
                 }
             }
         }
